*** empty log message ***
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@3355 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/x_mat.cpp
@@ -41,7 +41,7 @@ const M2_string IM2_Matrix_to_string(const Matrix *M)
   return o.to_string();
 }
 
-int IM2_Matrix_hash(const Matrix *M)
+unsigned long IM2_Matrix_hash(const Matrix *M)
 {
   return M->get_hash_value();
 }
